refactor: All MiniF2F into its own directory
