+++ experiments/minif2f/MiniF2F/.gitignore
@@ -1,0 +1,3 @@
+/build
+/lakefile.olean
+/lake-packages/*
