+++ experiments/minif2f/MiniF2F/lakefile.lean
@@ -1,0 +1,12 @@
+import Lake
+open Lake DSL
+
+require aesop from git
+  "https://github.com/leanprover-community/aesop.git" @ "v4.10.0-rc1"
+require mathlib from git
+  "https://github.com/leanprover-community/mathlib4.git" @ "v4.10.0-rc1"
+
+package MiniF2F
+
+@[default_target]
+lean_lib MiniF2F
