+++ experiments/minif2f/MiniF2F/lean-toolchain
@@ -1,0 +1,1 @@
+../../../src/lean-toolchain
