+++ experiments/minif2f/README.md
@@ -5,3 +5,12 @@ This is an experiment on running a LLM prover on miniF2F data. Run with
 ```sh
 python3 experiments/minif2f/main.py [--dry-run]
 ```
+
+## Developing
+
+Run unit tests with
+
+``` sh
+python3 -m model.{llm_agent,gen_tactic}
+```
+
