+++ experiments/minif2f/main.py
@@ -5,15 +5,17 @@ from typing import Optional
 from pathlib import Path
 from pantograph.server import Server, ServerError
 from pantograph.search import SearchResult
-from pantograph.search_llm import LLMAgent
+from model.llm_agent import LLMAgent
+
+PATH_EXPERIMENT = Path(__file__).parent.resolve()
 
 def get_project_and_lean_path():
-    cwd = Path(__file__).parent.resolve() / 'Example'
+    cwd = PATH_EXPERIMENT / 'MiniF2F'
     p = subprocess.check_output(['lake', 'env', 'printenv', 'LEAN_PATH'], cwd=cwd)
     return cwd, p
 
 def read_test_data(use_valid: bool):
-    jsonl_path = Path(__file__).parent / ('valid.jsonl' if use_valid else 'test.jsonl')
+    jsonl_path = PATH_EXPERIMENT / ('valid.jsonl' if use_valid else 'test.jsonl')
     with open(jsonl_path, 'r') as f:
         return [json.loads(l) for l in list(f)]
 
@@ -44,7 +46,7 @@ def output_file_name(datum, use_hammer: bool, use_llm: bool):
         folder += '-hammer'
     if use_llm:
         folder += '-llm'
-    folder = Path(__file__).parent / folder
+    folder = PATH_EXPERIMENT / folder
     folder.mkdir(exist_ok=True, parents=True)
     return folder / f"{name}.json"
 
@@ -65,7 +67,7 @@ def run_eval(args):
         if file_name.is_file():
             print(f"Skipping {datum['id']}")
             continue
-        server = Server(imports=["Example"], project_path=project_path, lean_path=lean_path, options=["maxHeartbeats=0"])
+        server = Server(imports=["MiniF2F"], project_path=project_path, lean_path=lean_path, options=["maxHeartbeats=0"])
         agent = LLMAgent(server, use_hammer=args.use_hammer, use_llm=args.use_llm)
         result = try_test_data(server, agent, datum, max_steps=args.max_steps, max_trials_per_goal=args.max_trials_per_goal)
         if result is None:
