+++ pyproject.toml
@@ -7,12 +7,8 @@ license = "GPL-3"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-# vllm = "0.4.1"
-numpy = "^1.26.4"
 pexpect = "^4.9.0"
 python = "^3.10"
-sglang = "^0.1.16"
-torch = "2.2.1"
 
 [tool.poetry.build]
 generate-setup-file = false
@@ -22,10 +18,14 @@ script = "build.py"
 # Experiment related dependencies here to not clutter the main project dependencies.
 fire = "0.6.0"
 notebook = "^7.2.1"
+numpy = "^1.26.4"
 openai = "^1.31.0"
+sglang = "^0.1.16"
 tenacity = "8.3.0"
 tiktoken = "^0.7.0"
+torch = "2.2.1"
 wandb = "0.17.0"
+# vllm = "0.4.1"
 
 [build-system]
 requires = ["poetry-core"]
